requirements-desktop.txt
pywebview>=5.2,<6.0
pyinstaller>=6.12,<7.0
